requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
ratelimit==2.2.1
diskcache==5.6.3
rich==13.7.0